update 6/13/2024
feat: Add include paths for Renesas Uno WiFiR4 variant

This commit adds additional include paths for the Renesas Uno WiFiR4 variant in the `.vscode/c_cpp_properties.json` file. These include paths are required for the proper compilation of the code.

In `webpage.h`, the water level is now displayed as a number without units. The `fetch('/rain')` function has been updated to set the `innerText` of the `waterLevel` element to the `data.waterLevel` value received from the server.

This commit improves the code by providing necessary include paths and updating the water level display in the webpage.